task_project/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
pytest-django>=4.5
pytest>=7.2
requests>=2.31

